+++ example/ca-census/index.ts
@@ -37,7 +37,7 @@ async function main() {
         const procInfo: { processId: string, processMetadata: ProcessMetadata } = JSON.parse(readFileSync(config.processInfoFilePath).toString())
         processId = procInfo.processId
         processMetadata = procInfo.processMetadata
-        processParams = await VotingApi.getProcessParameters(processId, gwPool)
+        processParams = await VotingApi.getProcessContractParameters(processId, gwPool)
 
         assert(processId)
         assert(processMetadata)
@@ -178,9 +178,7 @@ async function launchNewVote() {
     assert(entityMetaPost)
 
     // Reading back
-    const processInfo = await VotingApi.getProcess(processId, pool)
-    processParams = processInfo.parameters
-    processMetadata = processInfo.metadata
+    processParams = await VotingApi.getProcessContractParameters(processId, pool)
     assert.strictEqual(processParams.entityAddress, entityAddr)
     assert.strictEqual(processParams.startBlock, processParamsPre.startBlock, "SENT " + JSON.stringify(processParamsPre) + " GOT " + JSON.stringify(processParams))
     assert.strictEqual(processParams.blockCount, processParamsPre.blockCount)
@@ -390,7 +388,7 @@ async function checkVoteResults() {
         const envelopeHeight = await VotingApi.getEnvelopeHeight(processId, pool)
         assert.strictEqual(envelopeHeight, config.numAccounts)
 
-        processParams = await VotingApi.getProcessParameters(processId, pool)
+        processParams = await VotingApi.getProcessContractParameters(processId, pool)
 
         if (!processParams.status.isEnded) {
             console.log("Ending the process", processId)