+++ example/off-chain-census/index.ts
@@ -52,7 +52,7 @@ async function main() {
         const procInfo: { processId: string, processMetadata: ProcessMetadata } = JSON.parse(readFileSync(config.processInfoFilePath).toString())
         processId = procInfo.processId
         processMetadata = procInfo.processMetadata
-        processParams = await VotingApi.getProcessParameters(processId, gwPool)
+        processParams = await VotingApi.getProcessContractParameters(processId, gwPool)
 
         assert(processId)
         assert(processMetadata)
@@ -255,9 +255,7 @@ async function launchNewVote(censusRoot, censusUri) {
     assert(entityMetaPost)
 
     // Reading back
-    const procInfo = await VotingApi.getProcess(processId, pool)
-    processParams = procInfo.parameters
-    processMetadata = procInfo.metadata
+    processParams = await VotingApi.getProcessContractParameters(processId, pool)
     assert.strictEqual(processParams.entityAddress.toLowerCase(), entityAddr.toLowerCase())
     assert.strictEqual(processParams.startBlock, processParamsPre.startBlock, "SENT " + JSON.stringify(processParamsPre) + " GOT " + JSON.stringify(processParams))
     assert.strictEqual(processParams.blockCount, processParamsPre.blockCount)
@@ -352,7 +350,7 @@ async function checkVoteResults() {
         const envelopeHeight = await VotingApi.getEnvelopeHeight(processId, pool)
         assert.strictEqual(envelopeHeight, config.numAccounts)
 
-        processParams = await VotingApi.getProcessParameters(processId, pool)
+        processParams = await VotingApi.getProcessContractParameters(processId, pool)
 
         if (!processParams.status.isEnded) {
             console.log("Ending the process", processId)